Rewrite of the algorithm to be more intuitive.
Formerly, there was one transition method, transit(), which was either a consuming transmission (consuming a character from a tape, but emitting only an epsilon), or an epsilon transmission (not consuming anything from the tape, but emitting something).  There were various ways of checking which was which, like seeing if the tier or c was "__eps__", seeing if the query tiers were in the in tiers, etc.  (This was cumbersome, but back when a lot of these functions were more complex, it saved a lot of duplicated code.)

However, as we move away from the idea of "transducers" consuming and
emitting, and instead think of this in terms of a relational algebra (e.g.
thinking in terms of "joins" rather than "compositions"), it got rather
difficult to conceptualize how these ideas related to what was really
going on.  We've split the function back into two, roughly corresponding
to consuming and emitting transitions:

 - require(tier, target) gives the states that can be reached matching
   both tier and target.

 - query(tier, output) asks for any char and state that we can reach along
   the requested tier, then adds the tier:char to the output.

So for a join (probably the key operation in this whole deal), we start by
querying for one of the child states what the next possible chars are, on
a particular tier, and then requiring the other child to have it on the
appropriate tier.