Rewrite to deal with sequence problem
Background: Multi-tape systems typically still act like single-tape
systems underneath, just ones where each transition contains both a tape
label and a character.  The difficulty of multi-tape input with this trick
is the potential to receive content from different tapes in the "wrong
order", relative to what the network is expecting.  For example, we might
be joining ("composing", in the transducer metaphor) two transducers, the
first of which writes to tape A and then to tape B, and another transducer
that reads from tape B and only then from tape A.  This can lead to
blocking, even though tapes A and B are really unordered with respect to
each other.

So we want to be able to construct networks that can either (a) write to
any tape at any time or (b) read from any tape at any time.  (Either one
of those would be good enough.)  But we don't want to explicitly create
the network that has all the transitions such that you could follow any
one any time, that network would be huge.  (For example, for an otherwise
completely linear acceptor representing two literal strings on different
tapes, the automaton would be O(n*m).)

The previous commit proposed a solution for this, using a SMSequenceState
that effectively kept a different pointer to the same SequenceComponent,
but updating it separately for each different tape (i.e. stepping through
only those elements that are relevant for that tape).  That it to say, it
treats the result as the product of the projections (in the relational
algebra sense) of the sequence to each tape label.  But that's only true the
elements of the sequence have disjoint tape labels (like the really simple
example given above).

Rather than revert (because I liked a lot of the other changes I had
made, and couldn't easily disentangle them with the erroneous changes), I
just rewrote the whole thing again.  This time sequences can
rearrange elements according to what tape is being required.  For example,
consider a sequence P+Q, where P is a literal reading from tape a.  If we
call require("a", "x") on this sequence, and P really does begin with "x",
we return Rest(P)+Q as the next node.  However, if we call
require("b","x"), then we skip over P, and forward the requirement to Q.
If it's successful, we then return P+Rest(Q) as the next node.

Other changes:

(1) Projection (in the relational algebra sense) is implemented.

(2) Query no longer takes a tier argument, it's back to meaning "give me
whatever comes next, regardless of what tier it's on.  This was something
introduced to support the semantics of the erroneous Sequence anyway, and
it was messing with the semantics of projection.

(3) There is no longer a distinction between Components and States,
they're all the same thing now.  They already mirrored each other (like
Components with n children typically corresponded to States with n
children), States kept references to their corresponding components, and
Component.firstState() increasingly just called new State() for the
corresponding State.  So now instead of building a Component and later
asking it for a first State, we just build that first State at the outset.

(4) EmbedState just yields its child, it doesn't wrap itself around the
child state anymore.  This may come back to bite us, but we can always
change it back.