Added recursion protection
In previous versions of this engine, recursion worked, but would run into
the predictable troubles:

(1) in the case of right-recursion, emitting forever.  (Which is correct
behavior, but in practice confusing to the novice linguist-programmer.)

(2) in the case of left-recursion, blowing the stack.

So we need some kind of stack that the program can consult to see how deep
it is.

As predicted in the previous commit message, the previous implementation
of EmbedState messed with this.  Previously, EmbedState's successor was
just the successor of the child (that is, the successor of the grammar to
which the symbol refers).  I had been hoping not to have to wrap every
successor of an EmbedState's child in a new EmbedState.  But this doesn't
work when the EmbedState is the right child of a ConcatState, because the
successor of the ConcatState ends up with its right child *not* being an
EmbedState, and then when it's time to move on to that state it doesn't
have recursion protection because only EmbedStates have recursion
protection.

So now, of course, we restore the even older behavior of EmbedStates, that
of wrapping child successors in new EmbedStates.

Recursion protection currently involves passing around a Counter of symbol
names; when an EmbedState executes it checks whether its name is in the
Counter too many times.  If so, it fails; if not, it adds its name to the
counter and forwards on to the child.

In the future we can probably use the Output to keep track of this; we're
already passing that around.  (That is to say, we can say there's a
special tape for each symbol name, and then do operations on that tier to
maintain the corresponding symbol counter.)