Added "rename" operation
In the sense of Codd's (1969,1970) relational algebra.

Along with the join and project operations, this completes a relational
algebra over multi-tape grammars, at least with respect to the equality
relationship.  (That is to say, we haven't implemented theta-join for
relations other than equality, meaning we can't handle queries akin to
SQL's "select * where cost < 100".)