Any character parsers (i.e. "dot") + refactoring
The addition of the dot (e.g. h.llo matching "hello" or "hallo" or...)
ended up violating a key assumption of the previous algorithm,
necessitating changes throughout.

As noted in previous commit messages, this algorithm has gone back and
forth between having two different querying functions (one for a specific
tier/char, and one for a more general "give me what you've got"), and
treating those both as the same function that varies in behavior depending
on what it sees in the tier and target params.  Recently, we were going
with two, because of an invariance in the algorithm: a gimme-what-you-got
query (almost) always added something to the output, and a more specific
tier/target requirement never added something to the output.  So it was
awkward to treat them as one function with the same signature.

The dot ends up violating this idea.  While we could treat the dot as just
emitting anything in the entire vocabulary, one after another, we really
don't want the dot emitting, in turn, every character in the vocabulary.
Like picture doing a join between  ".ello" and "hello".  It's silly to
emit "a" and check whether "hello" can accept it, then "b", then "c"...
there's only one thing that's going to work.  It'd be better to just emit
a sort of "__ANY__" character and let the "hello" say "Yes, I accept that,
in my case that's an "h".  But this is a case of a require(tier, target)
type call on "hello", but it's one that DOES have an output, "h".  So
require(tier, target) needs to have an output again.

There are other future queries that are likely to have a similar effect,
like if you treat weights as their own tier, the actual output depends on
both of them.

With this change, require() and query() became so similar again that it
was more concise to collapse them back into a single function, now called
probe().

Another simplifying change is that we're now consistent in that any State
that gets an irrelevant query (i.e. not on a tier that it processes)
returns itself (rather than, say, failing).  This means that States no
longer have to advertise what they're willing to process, allowing us to
dispense with the logic for calculating tier relevance.  That was always a
little annoying to calculate because of EmbedStates potentially referring
to themselves.

Finally, we no longer pass around outputs when probing.  Anything that we
care about (that is, anything that's not Projection-ed away) ends up in
the final (tier, target, state) pair that breadthFirst() and depthFirst()
end up seeing.  So rather than make States accumulate the output during
probe(), the search algorithms do it at the end of each cycle.  (Someday
States might need access to that and we may have to start passing it
around again -- I can picture equivalents of our old "before" and "after"
transducers potentially needing it -- but we don't need to pass around
outputs to *construct* them anymore.)

These three changes end up making the overall algorithm somewhat shorter
and simpler; hopefully that simplicity remains as we further develop this.