Implementing repetition (e.g. star, plus)
As has become traditional, new additions break invariants in the algorithm
and require some rewriting.

Previously, one invariant had been that no states that are accepting() had
valid successors.  (That is to say, there were never any edges out of an
accepting state.)  That's true for, say, a Literal, but it's not true for
a Repetition.  If I've just finished repetition 2, but I'm allowed
anywhere between 1 and 10, then I'm in an accepting state, but I could
also transition to the start of repetition 3.

Softening this requirement required some algorithmic changes, because
ConcatState, on seeing that its left child is accepting(), moved on to
its right child.  Now, to handle a RepetitionState in the left child, it *both*
moves on its right child, and continues to attempt to probe() the left
child.

Doing so actually broke a number of tests, and in the end it exposed that
I had made some conceptual errors in other parts of ConcatState.  In
particular, we didn't have an overt signal anymore for "This tier is
irrelevant to me"; if a state received irrelevant input, it just returned
itself (or we could say "remains in place").  That made some things, like
JoinState.probe(), a lot simpler.  However, it doesn't work, in general,
to test whether an input was irrelevant by testing a child's result for
equality with that child.  A ConcatState can transition to
its child2 even with an irrelevant input, and that child2 isn't equal to
the original.  Anyway, it was necessary to reinstate an explicit signal of
whether a state had matched its input or whether it had ignored it,
specifically by returning a boolean among the many outputs of probe().