Implementing negation
Negating an automaton consists of toggling the accepting-ness of each
state (so that it fails exactly in those states where it would previously
have succeeded), and also adding a new success state to which we
transition when we cannot match from the current state.

Implementing the label for the transition to that state means being able
to say "Not any character that we could match by other transitions", e.g.
if we could transition out by "a" or "b" we have to be able to specify
that the additional edge is "Not(a or b)".  And since you can negate
anything, you can end up with a whole algebra of Not(Not(a) or b) sort of
things.  At one point I had started to implement that, but it gets
tedious.

So instead, characters are now bit vectors, so that we can do complements,
differences, any, none, etc. exactly when we need them.  This requires
tracking vocabs, so that we can set the correct indices in the bit
vectors.  This isn't too tricky, except for the Rename operator, since
this will change what different operators refer to particular tapes by.