Negation requires determination
The negation operator in the previous commit leads to some big trouble in
two cases:

(1) Where there's a non-determinism at the current state; it could go down
the "wrong" one and end up accepting a join with something that would have
violated the other path.

(2) When the child state can be in an infinite spiral of acceptance (like
.*).  The negation state will thus be in an infinite spiral of
non-acceptance (i.e., it will never accept along that spiral, so it will
keep traversing forever), meaning there can be things like languages with
0 sentences but we'll keep churning forever looking for them.

I don't know how to solve the second without actually making the entire
graph and finding loose ends.  (It's really an instance of a larger
problem with doing these things on the fly rather than building the
graph.)  But the first we can do piecemeal: just looking at the various possible next states and rearranging them when necessary to make the choice deterministic.

So the old "probe" function is now called "ndQuery" (for
non-deterministic), and there's now a "dQuery" function that just calls
ndQuery, makes the results disjoint, and yields those.  Strictly speaking
we only need to call dQuery from NegationState, but we're probably better
off overall always calling it, because it keeps UnionState from exploding
the number of states in the search agenda.