Tape now an object, subsumes Vocab
There was an annoying redundancy in the way tapes and vocabs were handled.
Formerly, tapes were just strings (naming the tape), and RenameState would
just pass in a different string to its children.  However, it also had to
wrap the Vocab so that its children, upon using that different tape name,
would actually get the correct tape for the purpose of indexing characters
correctly.  This was just convoluted.

Now, it's a Tape object's responsibility for knowing its own name and also
what characters are assigned to what indices.  That way all RenameState
does is wrap the tape object itself, with a RenamedTape, which handles
both of those.

It's still a bit convoluted due to the possibility that a RenamedTape
might hold the "Any" tape, the one that contains all the tapes.  But it's
less convoluted than the original.