Refactoring Outputs
The previous implemenation of Outputs had all the tapes interleaved, like

   [t1:h t1:e t2:w t1:l t1:l t2:o t1:o ...]

This was easy but consider "flag"-type tokens.  Flags are pretty much
ordinary strings except that they fail if they're added to a string that
doesn't match them.  When we're interleaving like the above, however,
we've got to search back in the string to find the most recent content on
that tape.

So now we store them in a non-interleaved format, like

   t1: [h e l l o ...]
   t2: [w o ... ]

That way we can instantaneously fail when something inappropriate is
added.

This isn't a big gain time-wise, but it'll make things easier in the
future when we start to reintegrate Outputs and Tapes, I think.