On-demand loading of multiple-source projects
Projects need to be able to be spread over multiple sheets, but we don't
want to necessarily load and compile the entire spreadsheet/directory
pre-emptively.  This latest commit makes it so that the client only loads
one source file, and then we load other source files in the same directory
only if they're necessary to resolve the symbol names in already-loaded
files.