Integration of new language with Google Sheets
This update integrates the core of the old Google Sheets add-on with the
new language.  I had hoped that it'd be a simple drop-in replacement, but
(a) the APIs are sufficiently different now, and (2) the Google Apps
Script transpiler and TypeScript were not playing nice together, requiring
a new way of packaging up the core Gramble.  (Google Apps Script basically
has no module system, and I couldn't find a way to make both the
TypeScript compilation happy and also have things load correctly within
GSuite once they've been compiled to GAS.  So now we browserify Gramble
like we would for embedding in a browser, and just write the Sheets add-on
in JavaScript.)

Right now, the only functionality exposed in the Sheets add-on is
highlighting, but it at least shows that Gramble is loading properly
within GSuite.