Re-integrating with Google Sheets
The parsing engine had changed enough that getting it to work with the old
Google Sheets add-on wasn't trivial.  Most of Code.js and sidebar.html had
to be re-written.

The new Sheets interface concentrates on the sidebar, rather than menu
options; the only menu options are now highlighting and showing the
sidebar.

Also in this commit are some (relatively minor) changes to the parsing
engine to support interactions that the sidebar needs, like reporting the
tape names/colors for any named symbol, sampling randomly from a symbol,
and parsing (that is, joining with) an arbitrary user-supplied string
dictionary.