Fixing major bug with ConcatState
We found a major (but stupid) bug in which you could match incomplete
material (e.g. matching "h" with "hello") so long as the "h" was
concatenated with material on another tier.  E.g.:

   (text:h + gloss:foo) & text:hello

would succeed with the result {text:hello, gloss:foo}.

The reason for this is that the text:h succeeds, and the ConcatState would
move on (meaning it would yield its child2 as its successor), so that the left side here is only gloss:foo.  This also
succeeds.  By the time we get to text:ello on the right side, there's
nothing to contradict it on the other side!

So the concatenation has to remember that, in the past, it cared about the
"text" tape, so that, if it receives "text" content in the future, it can
appropriately reject it.  So instead of yielding child2 as its successor,
it constructs a new ConcatState with a boolean child1Done = true.

This also requires that states know what tapes they care about.
Previously, at a probably-misguided attempt at simplicity, states didn't
actually keep track of that.  A state not caring about a tape was just
indicated by the state not going anywhere if you gave it material on that
tape.  This ends up making it difficult (if not impossible) to make some
of the decisions we need to in the new ConcatState etc.  So now, instead,
we calculate this when necessary, and (for those states where the
calculation isn't completely trivial) remember it and pass it on to
successor states so that we don't have to keep calculating it.