Implementing flags with joins
Users of XFST/LEXC are used to using "flags" to constrain the output of
the parse.  I typically use "U" (for "unify") flags; if the flag is not
yet set, it sets it, and if the flag IS already set, it checks for
equality.

However, we don't really need any extra mechanism to do this; our
algorithm is already a unification algorithm across any number of tapes.
The operation of setting something if unset, or checking for equality if
set, that's just "join" from our point of view.  That is to say, if you
have some symbol X and you want to "set a flag" so that only results from
X with that "flag" on a particular tape succeed, you can just say
join(tape:flag, X).

We just need a non-cumbersome tabular syntax to do this.  (Previously, you
could only use the join command, which applied to the whole previous
*table*, whereas a lot of times you want a flag to only apply to a
previous *line*.)  So now, simple joins can be expressed by a header
"@t", like

   verb, table, @obj, embed, embed, text, gloss
                2SG, root, suffix, foo, IMPER

which would constrain this line to only producing results where the obj
tape is equal to "2SG" (or, where there is no obj tape at all).

We do this by saying that each kind of header also defines a merge
operation (usually concatenation), and then recursively merge the
resulting states right-to-left.  (So text:foo and gloss:IMPER are
concatenated, then embed:suffix is concatenated with the result of that,
etc.)  A new "FlagHeader" is the only exception to this concatenation;
instead of concatenating, it joins (so obj:2SG is joined with the result
of everything to its right).

Special care has to be taken for slash headers here, because it's probably
going to be very common for programmers to want to slash-concatenate text
or a flag, or two flags.  Previously, slash headers just concatenated
the relevant child states, but that would end up just ignoring the
join-i-ness of mood in "gloss/@mood".  So instead of concatenating,
slashes now just compile-and-merge each one in turn, going right to left.
(So in this example, merging @mood with everything to the right, and then
merging gloss with the result of that.)