Refactored the stateMachine tests to use new testGrammar function.
Added testMatchOutputs() and testGrammar() functions to testUtils.ts.

testMatchOutputs checks that outputs as produced from State.generate() match the
expected outputs. The order of the outputs does not matter. testMatchOutputs
just checks that all the required dictionaries and only those dictionaries are
in the outputs list.

testGrammar call State.generate for a grammar, and then checks the results.
It is coded in such a way as to catch any unexpected exception and report it in
a specially constructed mocha/chai test so as not to blow up the VSCode's Mocha
Test Explorer during test discovery.

Some additional tests were added to some of the test suites.

In particular, three tests were added to testNegation.ts that fail. These
tests reveal deficiencies in the NegationState code that need to be addressed.

testSheetParser.ts has not yet been updated to use testMatchOutputs.