Fixed some (but not all) negation bugs
Fixed all but one of the negation bugs, but in the process found a bug in
rename.

The negation bugs were due to two things:

1. The issue with ANY_TAPE getting added to a tape collection stemmed from
a deeper conceptual problem, that the implementation was only really
thinking about the negation of languages on a single tape, and thus only
considered the complement of the character on that tape.  (That is, if the
vocab on tape T1 is {h,e,l,o}, then ~h would be {e,l,o}.)  But really it's
not, it's any character on any tape.  We can't represent the emission of
that with our current organization of tokens, tapes, and vocabs; we'd need
to change how we do them so that the negation of T1:h would also
potentially include characters on other tapes.  But I'm not sure it's
currently worth it to make that change -- it's a lot of work, it probably
has some unintended consequences, and I'm not sure we actually *need* it
for anything, because all the applications of negation that I was sure we
needed were negations of a single tape.  Anyway, so instead, I just made
it a rule that negation can only scope over single-tape grammars, and thus
negation never has to emit on an ANY_TAPE, it always knows what to emit
on.

2. There was also a bug in the new tape-to-string implementation; I just
put a key operation on the wrong level of nested looping, which was giving
incorrect string outputs for tape outputs that otherwise would have been
correct.

There's a remaining bug, that I haven't gotten to yet.  The issue
is when you join(A, B) where A will emit infinitely.  The join algorithm
only ever emits from B (and checks with A) when A doesn't emit anything.
(Otherwise there are many "diamond" cases where we end up with the same
result from two different paths.)  But something like ~(text:x) can always
emit something, so we never get around to it.  This is true even if we
prevent infinite generation by putting a character cap: it keeps emitting
up to that limit, then fails if there's anything in B on an unrelated
tier, because it hasn't managed to get to that accepting state.

I don't know what's causing the bug in Rename state -- haven't looked into
it yet -- but I found it when playing around with reversing the orders of
joins.  That suggests something about rename/join is 'directional' in a
way that it shouldn't be.