Fixed remaining negation and rename bugs
Both ended up being easy.  (NB: The negation fix actually ended up in the
last commit by mistake, so look there if you want to see the diff.)

1. Negation

The problem was that an infinitely emitting state like ~(text:x), when on
the left side of a join, will never give the right hand a chance to emit,
even if the left hand is accepting, and thus we never get to unrelated
content on the right hand side, and thus never get to
accepting.

So now, when the left hand of a join is accepting, we first give the right
hand a chance to emit.  (It's just the same algorithm in reverse, giving
the right semijoin priority over the left semijoin.)

2. Rename bug

Very simply, this was because RenamedTape, when asked for its tapeName,
gave its child's name... even when it had renamed it!  So it caresAbout()
was making the wrong decision about whether it cared about the tape, and
this was messing up ConcatState.ndQuery().