In-document unit testing
Now you can add a "test" command below any operation (e.g. table, join),
with a table of tests, in order to test that each line of the tests can be
generated by the grammar above.  For example:

word	table	embed	embed
		verb	suffix

	test	text	gloss
		foobar	run-1SG
		moobaz	jump-2SG

These tests have no effects on the state generated, but lines that cannot
be generated will be highlighted with errors in a dev environment.

This was fairly simple to implement using two recent updates:

1. The syntax tree of Components, independent from the State tree itself,
on which we can run runChecks().

2. The operation [left] Semijoin.  Semijoin(A, B) requires of every result
b of B is a subrecord of some a of A (that every field in b has the same
value in a, but not necessarily the other way around).  We don't require
every field in a to have the same value in b because in practice it's hard
to keep track of every field that A can generate, there can be fields in A
that are necessary for a calculation but that the programmer doesn't care
about in every downstream test.