Reintegration with sheets add-on
The recent changes to the sheet parser neglected to highlight things
properly, which only became apparent in the Sheets interface.  Fixed that.

Also fixed a bug in the sidebar where the presence of a warning was short
circuiting the display of errors, because I was "returing" where I should
have been "continuing".

Also gave more informative "short messages" for some errors.