Simplified X/Y header coloring
Previously the color of an X/Y header was the HSV average of X and Y.
This was cute but ends up less readable overall; e.g. if "text" is always
green, adding "/root" to that in one table changes the color and takes
away that visual cue that it's the same field as other text fields.

So now the color of X/Y is just the color of X.  If the programmer would
rather it be the color of Y, well, they could call it Y/X instead.