Fixed major bug in randomization
Previously, random generation would never, if given a Union(A, Empty()),
would never generate A!  This is because Union(A, Empty()) is accepting,
so the algorithm results in a sucessful parse here without ever calling
ndQuery, randomization only ever returns one parse, and ndQuery is where
the randomization stuff happens.

So now, randomization doesn't happen in dQuery/ndQuery anymore, there's a
separate recursive setRandom(true) call that propagates through the tree.
When a UnionState receives this message, it chooses a random child to be
its proxy -- both for ndQuery and for accepting().  That way, if it
happens that A is the random child, Union(A, Empty()) ISN'T accepting, and
that short-circuiting doesn't occur.