Moved "flag" joins to be left associative
I've been pondering whether inline operators like "@" and the planned
inline phonological rules should be left-associative or right-associative.
(That is, whether they apply to everything to the left of them or
everything to the right of them.)  Before this they were
right-associative.

I think left-associative is easier for the novice linguist-programmer to
understand though, because I think they'll tend to understand these
grammars in terms of generation rather than parsing.  If you think in the
direction of parsing, you think of phonological rules being UNdone to the
parse input (so that it represents a more underlying form) before being
parsed by a grammar that represents those underlying forms.  If you think
in the direction of generation, you think of the generation of all the
underlying forms happening first, and then the phonological rules are
applied to them to generate the surface forms.  I think the latter is more
intuitive.

If the programmer is conceptualizing things as happening in a
left-to-right order (which they kind of have to, given that that's how
concatenation is ordered), that means rules should be to the right of what
they apply to.  This is especially noticeable when the programmer puts two
rules in the same line -- if rules were to be right-associative, they
would appear to be in the "wrong order" compared to how they're typically
conceptualized by linguists.

Likewise, if you're thinking of parsing, you may think of filtering a
table so that you only match, say, subj=1SG as something you do first.
That's part of your parse input, so you get things that match it by
putting it into the input, and that's a thing you do beforehand.  On the
other hand, if you're thinking of generation, you probably think of
filtering as something that happens after the fact.  You generate a bunch
of outputs (from nothing), then throw out everything that doesn't match
subj=1SG.  So I think having @subj:1SG coming after what it applies to is
more consistent with a mental model of this as generating.

It should be emphasized that neither of these choices affects the
implementation in either way, it's not like we're *really* generating all
the possibilities and then throwing them out.  In reality, these operators
surround what they apply to whether we decide to show them on the left or
the right.  It's just that putting them on the right (that is, making them
left-associative) is probably more in line with a naive mental model of
what the system is doing.