Major bug fix about renamed tokenization
While trying to implement the equivalent of templates in the old Gramble,
I encountered a bug due to my misconception about tokenization.

The problem comes down to:

1. Literals tokenize their text ahead of time, for convenience

2. However, that literal can potentially be reached by several possible
paths through RenameStates that rename that literal's tape!  For example,
if the literal is within a named symbol, and then two different
RenamedStates apply to EmbedStates referring to that symbol.  (This sounds
like a bizarre configuration but it does come up some templates.)

3. What ended up happening is that the literal could get tokenized several
times in the scope of different RenamedStates, and ends up with whatever
the last one was.  That leads to unexpected consequences in
parse/generation paths that don't go through that last RenameState, like
failing to produce output or producing garbled output.

So we have to go back to the LiteralState only generating tokens at parse
time, according to the actual tape it's reading/writing from.