Fixed another Rename bug
RenameStates were trying to unwrap the child's result tape even if they
hadn't renamed it in the first place (which happens when it's irrelevant
to them).  So we have to remember if we wrapped it and then undo it only
if we did.  I also put in a test for this.

Also made some changes to the Project API to make it easier for library
users to make simple projects.  You now don't have to instantiate a
DevEnvironment for whatever your environment is, in order to load
multi-sheet projects.  Now, if you construct a new Project() without
arguments, it instantiates a dummy DevEnvironment.  This environment can't
do recursive imports (after all, it doesn't know what your file
environment looks like), but you can add source files to it manually.

Note that Project no longer automatically runs runChecks() after
addSheet(), because it no longer knows that that's the last sheet.  You
have to run runChecks() manually if you want those checks to be done.