Simplified UnionState
When the union A|B|C|D is represented hierarchically A|(B|(C|D)), or the
other way around, this makes choosing a random child difficult.  If our
binary UnionState chooses between two children with equal probability,
that doesn't result in equal probability between A, B, C, and D the way
the programmer probably hopes.

Treating UnionState as having any number of children, rather than being
binary, makes this a lot simpler.

I also made a start on semi-compilation (like partially constructing the
state graph where it would do the most good, without having to construct
the whole thing) but it was much more difficult than anticipated, so I'm
going to leave my attempt there, but unused, and come back to it later.