Incremental compilation
This commit gets a version of incremental compilation working, not
necessarily the most efficient version but enough to pass our tests.
"Incremental" in the sense that we don't generate the state graph in its
entirety, we just create some of it where it matters most.  Here, we're
basically just compiling the first node in the graph for each State.
(Think, for example, of a UnionState with 1000 children.  It's only the
very first decision in that UnionState -- the one where you iterate
through 1000 children -- that contributes most to the time complexity of
the parse, by the time you're parsing the second letter you've probably
discarded 90+% of those.  So just compiling that very first step is
probably going to give you the lion's share of time improvement.)

Incremental compilation here basically just treats the outputs of dQuery
as if they're nodes/transitions.  (That, after all, is what dQuery is for,
to make the dynamic parse look enough like a DFSA that negation won't mess
up.)

The one thing that gets dicey here is vocabularies.  Compiled grammars
have to be run with the same TapeCollection they were compiled with.  (We
can't just read the vocabulary off the actual grammar, the way we do with
normal states.  You need a vocabulary to compile (because compiling is
just partially running the algorithm), and that's going to be the vocab of
the original state.  You can't throw that away, because it can't be
reconstructed on the compiled grammar, since that's a different grammar
and the vocabulary collection algorithm won't necessarily get the same
results.  So you've got to keep it around; we now store TapeCollections on
the states themselves.  (Not for every state, just for ones for which
vocabularies have been calculated or inherited.)

This can cause potential problems in things like parsing and testing.
Parsing and testing work by joining/semijoining with the existing grammar.
That's a new grammar, and it would ordinarily calculate a new vocabulary
before parsing.  But if it does that, you can run into problems, because
then it will be running one of its subgrammars -- the one that was
previously compiled -- with a new vocab.  So we need to make sure, in
these situations, that the vocab of the new grammar only adds vocab to the
existing grammar.  (E.g., if the parse or the test use a tape or token
that the existing grammar doesn't, that would introduce something new to
the vocab.  We have to make sure those are added to the existing vocabs,
so that it doesn't mess up the indexing of the existing tokens.)  Anyway,
that's a priority for the API, to make it less dangerous to do that, but I
haven't gotten around to it tet.