Better incremental compilation
This commit makes a number of improvements to incremental graph
compilation:

1. As mentioned in the last commit, it's possible to mess up parsing/unit
testing by trying to generate a compiled state with a fresh vocab, which
makes it possible for an application programmer to make mistakes in
parsing that would surprise them.  So now there are special methods in
State that handle this (parse() and runUnitTest()), rather than make the
application programmer do the right steps on their own.

2. You can now specify how many steps ahead you want to compile, with
compileLevel.  compileLevel=1 is precompiling only the first character, 2
the first two characters, etc.  So far 2 looks like a sweet spot.

3. In the previous commit, there was no special mechanism to deal with
EmbedStates compilation -- it was just whatever the run-time result of
EmbedState.dQuery would have been.  But this isn't great, because every
invocation of EmbedState.compile ends up with its own compilation.  (That
made recursion no problem, incidentally.  But we don't want thousands of
separate instances of compiled EmbedStates that are logically identical.)

In this commit, we make sure that every compilation of an EmbedState
refers to the same object (or, to be precise, every compilation at a given level of
recursion, we still want those to be different or else we mess up
recursion counting).  In order to do this, Namespaces have to be
compilation-aware.  On compilation, EmbedStates ask their Namespaces to
compile their referents; the Namespaces keep track of the details of this
so that EmbedStates don't have to know the details of where compiled
symbols are stored/named/accessed.  (Right now, they use the same
Namespace.symbol object that normal ones do, they just get a special
suffix on their name indicating that they were generated at a particular
level of recursion, like the suffix "@@@verb2" means that there were 2
instances of "verb" on the stack at compilation.)

I also did a bit of refactoring of the Namespace class to make some
aspects of this easier; in particular, the way Namespace.Symbol references
are resolved.

Oh, also, EmbedState is no longer a UnaryState.  Since it handled its
child differently, it actually didn't have any shared code with the other
descendents of UnaryState; may as well let it be its own thing to better
allow the other descendents to share code.