Small fixes & tests re: determinization
Playing around with removing/trivializing dQuery for particular classes
revealed some... not bugs exactly, but places where the algorithm would be
more robust to future change if we make some small changes now.

1) TextState (the abstract ancestor of Literal and AnyChar) used to check
whether its output was empty in ndQuery().  I took that out because dQuery() checks for
that.  But the dQuery algorithm on these is trivial, its only effect on
these is doing that check.  That's too close to relying on an undocumented
side effect for me, so back in ndQuery the check goes.  That way if we
change how/where dQuery applies, TextState will work regardless.

2) When dQuery is removed from ConcatState, we run into some double-output
problems (where the same output is returned multiple times), due to
yielding from the "Child 1 is done, move on, query child 2" block and the "Child 2 cares
about this, query child 2" block.  Only one of these should yield, so out
comes our old standby: var yieldedAlready.  (It's not clear to me how
dQuery was eliminating this bug; I would think it would preserve it.)

Regardless, we can't eliminate dQuery from ConcatState; ConcatState is one
of the states that is genuinely nondeterministic and so we need to keep
that around.  But like the above, I would rather ndQuery actually work on
its own, rather than rely on a side effect of dQuery.  So back in ndQuery
the yieldedAlready goes.

3. Added some further tests for Negation, especially one (~(h*ello) &
hello) that will fail if ConcatState isn't properly determinized.